+++ src/main/java/fsm/service/TableService.java
@@ -4,12 +4,13 @@ import fsm.domain.Table;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Set;
 
 public interface TableService {
 
 	public Integer addTable(Table table);
 
-	public Integer addAllTables(List<Table> table);
+	public Integer addAllTables(Set<Table> table);
 
 	public void removeTable(int tableId);
 
@@ -17,6 +18,6 @@ public interface TableService {
 
 	public Table getTableById(int tableId);
 
-	public List<Table> getAllTables();
+	public Set<Table> getAllTables();
 
 }
