+++ src/main/webapp/WEB-INF/applicationContext.xml
@@ -50,9 +50,9 @@
       <props>
         <prop key="hibernate.dialect">${jdbc.dialect}</prop>
         <prop key="hibernate.show_sql">true</prop>
+          <prop key="hibernate.hbm2ddl.auto">validate</prop>
       </props>
     </property>
-
     <property name="annotatedClasses">
       <list>
         <value>fsm.domain.City</value>
